backend/requirements.txt
fastapi
uvicorn
langchain[google-genai]
python-dotenv
sqlalchemy
alembic
asyncpg
psycopg2-binary
bcrypt
python-jose[cryptography]
